Fixed type

+++ contrib/activity/file/src/main/scala/com/krux/hyperion/contrib/activity/file/FileSplitter.scala
@@ -10,7 +10,7 @@ case class FileSplitter(
   header: Option[String],
   numberOfLinesPerFile: Long = Long.MaxValue,
   numberOfBytesPerFile: Long = Long.MaxValue,
-  bufferSize: Int,
+  bufferSize: Long,
   compressed: Boolean,
   temporaryDirectory: File
 ) {
